AI/loanmaker-ai/test_models.py
import os
from dotenv import load_dotenv
from google import genai

load_dotenv()  # <-- REQUIRED

api_key = os.getenv("LLM_API_KEY")

if not api_key:
    raise ValueError("LLM_API_KEY not found in environment")

client = genai.Client(api_key=api_key)

for model in client.models.list():
    print(model.name)