AI/loanmaker-ai/requirements.txt
fastapi
uvicorn[standard]
numpy
pandas
scikit-learn
joblib
pydantic
pytest
qdrant-client
openai
mysql-connector-python
python-dotenv
